website/requirements.txt
biopython
django
pandas
requests